app: Convert color_type to signed int so make the check meaningful
This fixes a GNU Clang warning:

warning: result of comparison of unsigned enum expression >= 0 is always true

+++ app/actions/gradient-editor-commands.c
@@ -76,7 +76,7 @@ gradient_editor_left_color_type_cmd_callback (GimpAction *action,
   gimp_gradient_editor_get_selection (editor, &gradient, &left, NULL);
 
   if (gradient        &&
-      color_type >= 0 &&
+      (gint) color_type >= 0 &&
       color_type !=
       gimp_gradient_segment_get_left_color_type (gradient, left))
     {
@@ -197,7 +197,7 @@ gradient_editor_right_color_type_cmd_callback (GimpAction *action,
   gimp_gradient_editor_get_selection (editor, &gradient, NULL, &right);
 
   if (gradient        &&
-      color_type >= 0 &&
+      (gint) color_type >= 0 &&
       color_type !=
       gimp_gradient_segment_get_right_color_type (gradient, right))
     {
